feat: make admin screens web-only, removing email collection from store binaries
Play's Data Safety check flagged the app for email collection: the admin
login (Supabase email+password auth) shipped in the native JS bundle even
though it's staff-only with no visible entry point. Rather than declare
collection for a feature end users never touch, the admin surface is now
genuinely web-only:

- real screens moved to src/features/admin/*.web.tsx (imports adjusted)
- native builds resolve the sibling *.tsx stubs, which render an
  "admin tools are on the web" notice (AdminWebOnly)
- app/admin/* route files are thin re-exports, so routes still resolve
  on every platform and the web app keeps full admin functionality

Platform selection happens at Metro module resolution, not in the router:
renaming route files to .web.tsx directly does NOT work (expo-router's
route context bundles every file in app/ regardless of platform suffix)
- verified empirically by exporting bundles and grepping for admin
markers. Android bundle now contains no manage-admin-users /
admin_audit_log / app_admins / sign-in strings; the web bundle retains
them all. Data Safety can therefore honestly declare no email collection.
