+++ DEPLOYMENT_GUIDE.md
@@ -118,4 +118,5 @@ eas update --branch production --message "Fix: summary persistence and UI refine
     *   **Build counters** (`versionCode`/`buildNumber`) are managed remotely by EAS (`cli.appVersionSource: "remote"` in `eas.json`) and auto-increment on every production build. Do not add these fields to `app.json`; if a counter ever needs manual correction, use `eas build:version:set`.
     *   **User-facing version** (e.g. `1.5.0`) is `expo.version` in `mobile-app/app.json` — bump it there for each release.
 *   **Sentry**: `EXPO_PUBLIC_SENTRY_DSN` (EAS Environment Variables, Production, Plain text) enables error/crash reporting; `SENTRY_AUTH_TOKEN` (EAS project secret) authorizes the `@sentry/react-native/expo` plugin's build-time upload (JS source maps always; R8 `mapping.txt` too, via `experimental_android.enableAndroidGradlePlugin` in `app.json`). **Both must exist as EAS values before building** — a production build with the plugin present but no `SENTRY_AUTH_TOKEN` fails outright (the upload is a build-graph-finalizing step on both platforms, not an optional best-effort step). `development`/`preview` profiles set `SENTRY_DISABLE_AUTO_UPLOAD=true` in `eas.json` since they have no token. Play Console crash reports are unaffected either way (the R8 mapping is embedded in the AAB).
+*   **Play Data Safety — email collection**: the store binaries collect **no** email addresses. Staff/admin login (Supabase email+password) is web-only: the real screens live in `src/features/admin/*.web.tsx` and Metro bundles the native stubs (`*.tsx` → `AdminWebOnly`) into iOS/Android instead. Keep it that way — moving admin code back into a shared path would re-trigger the Data Safety mismatch flag. Verify after native-facing changes with `npx expo export --platform android` and grep the bundle for `manage-admin-users` (must be absent).
 *   **Before the SDK 54 upgrade** (required for Play's API 36 deadline, ~Aug 2026): audit tablet/foldable layouts — Android 16 ignores the portrait orientation lock on large screens, so every screen must tolerate landscape/resized windows.
