+++ mobile-app/src/features/admin/account.tsx
@@ -1,0 +1,1 @@
+export { default } from "../../components/AdminWebOnly";
