+++ mobile-app/src/features/admin/login.tsx
@@ -1,0 +1,1 @@
+export { default } from "../../components/AdminWebOnly";
