+++ mobile-app/src/features/admin/logs.tsx
@@ -1,0 +1,1 @@
+export { default } from "../../components/AdminWebOnly";
